arreglo de estilo del fieldset
