+++ routes/web.php
@@ -18,6 +18,7 @@ return view('theme\layout');
 Route::get('inicio/Layot','administracion\Layoutcontroller@index')->name('layout');
 Route::get('/identificacion','identificacion\datosgeneralescontroller@index')->name('datos_generales');
 Route::get('docente/identificacion','identificacion\docentecontroller@index')->name('docente');
+Route::get('estudiante/identificacion','identificacion\estudiantecontroller@index')->name('estudiante');
 Route::get('administrativo/identificacion','identificacion\administrativocontroller@index')->name('administrativo');
 
 /*Administrador*/
